% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adaptive_method_match}
\alias{adaptive_method_match}
\title{Match a specified}
\usage{
adaptive_method_match(method, ..., error_call = caller_env())
}
\arguments{
\item{method}{Name of a Workday Adaptive Planning API Method.}
}
\description{
Match a specified
}
\keyword{internal}
