% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adaptive_method_template}
\alias{adaptive_method_template}
\title{Get a Workday Adapting Planning API template for a method}
\usage{
adaptive_method_template(method, ...)
}
\arguments{
\item{method}{Name of a Workday Adaptive Planning API Method.}
}
\description{
Use \code{\link[glue:glue]{glue::glue_data()}} to fill a template for a specified method with the
supplied parameters.
}
\keyword{internal}
