% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{adaptive_req_method}
\alias{adaptive_req_method}
\title{Create a Workday Adaptive Planning API request with a specified method}
\usage{
adaptive_req_method(
  req = NULL,
  method,
  caller_name = getOption("adaptiveapi.caller_name", "adaptiveapi"),
  username = Sys.getenv("ADAPTIVE_USERNAME"),
  password = Sys.getenv("ADAPTIVE_PASSWORD"),
  key = "ADAPTIVE_KEY",
  ...,
  version = getOption("adaptiveapi.version", 39)
)
}
\arguments{
\item{method}{Name of a Workday Adaptive Planning API Method.}

\item{...}{Additional named parameters passed to \code{\link[=adaptive_method_template]{adaptive_method_template()}}}

\item{version}{Workday Adaptive Planning API version. Defaults to
\code{getOption("adaptiveapi.version", 39)}}
}
\description{
Use \code{\link[=adaptive_request]{adaptive_request()}}
}
