% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{adaptive_request}
\alias{adaptive_request}
\title{Create a Workday Adaptive Planning API request}
\usage{
adaptive_request(
  base_url = "https://api.adaptiveplanning.com/api/",
  version = getOption("adaptiveapi.version", 39)
)
}
\arguments{
\item{base_url}{Base URL for the Workday Adaptive Planning API call. Defaults
to \code{"https://api.adaptiveplanning.com/api/"}}

\item{version}{Workday Adaptive Planning API version. Defaults to
\code{getOption("adaptiveapi.version", 39)}}
}
\description{
Create a Workday Adaptive Planning API request
}
\seealso{
\code{\link[httr2:request]{httr2::request()}}
}
