% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{req_body_xml}
\alias{req_body_xml}
\title{Add an XML file to a request body}
\usage{
req_body_xml(req, data = NULL, ..., file = NULL, overwrite = TRUE, type = NULL)
}
\arguments{
\item{data}{An \code{xml_document}, a list coercible to an \code{xml_document}, or a
character string or other object readable with \code{\link[xml2:read_xml]{xml2::read_xml()}}. Optional
if \code{path} is supplied.}

\item{path}{Optional if data is supplied.}
}
\description{
Add an XML file to a request body
}
\seealso{
\code{\link[httr2:req_body]{httr2::req_body_file()}}
}
